+++ lib/src/main/java/com/smileidentity/SmileID.kt
@@ -26,6 +26,7 @@ import okhttp3.logging.HttpLoggingInterceptor
 import retrofit2.Retrofit
 import retrofit2.converter.moshi.MoshiConverterFactory
 import timber.log.Timber
+import java.net.URL
 import java.util.concurrent.TimeUnit
 
 @Suppress("unused")
@@ -40,6 +41,9 @@ object SmileID {
 
     // Can't use lateinit on primitives, this default will be overwritten as soon as init is called
     var useSandbox: Boolean = true
+        private set
+
+    var callbackUrl: String = ""
         private set
 
     internal var apiKey: String? = null
@@ -133,6 +137,19 @@ object SmileID {
         val url = if (useSandbox) config.sandboxBaseUrl else config.prodBaseUrl
         retrofit = retrofit.newBuilder().baseUrl(url).build()
         api = retrofit.create(SmileIDService::class.java)
+    }
+
+    /**
+     * The callback mechanism allows for asynchronous job requests and responses.
+     * While the job_status API can be polled to get a result, a better method is to set up a
+     * callback url and let the system POST a JSON response.
+     *
+     * @param callbackUrl The callback url that will be used to asynchronously send results of your
+     * job requests
+     */
+    @JvmStatic
+    fun setCallbackUrl(callbackUrl: URL?) {
+        SmileID.callbackUrl = callbackUrl?.toString() ?: ""
     }
 
     /**