+++ lib/src/main/java/com/smileidentity/viewmodel/BiometricKycViewModel.kt
@@ -92,7 +92,7 @@ class BiometricKycViewModel(
             val authResponse = SmileID.api.authenticate(authRequest)
 
             val prepUploadRequest = PrepUploadRequest(
-                callbackUrl = "",
+                callbackUrl = SmileID.callbackUrl,
                 partnerParams = authResponse.partnerParams,
                 signature = authResponse.signature,
                 timestamp = authResponse.timestamp,